config/requirements.txt
Flask
Flask-CORS
mysql-connector-python
groq
python-dotenv
opencv-python
tensorflow
numpy
pandas
scikit-learn
pillow
requests
pydantic
psutil
pynput
PyAudio
mss
pyautogui
schedule
flask-sqlalchemy
pyjwt
bcrypt
python-dateutil
matplotlib
seaborn
scipy
joblib